fix: improve semantic of the HTML in main navigation
