+++ src/ui/components/Header.tsx
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import { Suspense } from "react";
-import { ActiveLink } from "./ActiveLink";
+import { NavLink } from "./NavLink";
 import { AccountLink } from "./AccountLink";
 import { CartNavItem } from "./CartNavItem";
 import { executeGraphQL } from "@/lib/graphql";
@@ -21,29 +21,29 @@ export async function Header() {
 							ACME
 						</Link>
 					</div>
-					<div className="flex overflow-x-auto overflow-y-hidden whitespace-nowrap lg:px-0">
-						<div className="flex gap-4 lg:gap-8">
-							<ActiveLink href="/products">All</ActiveLink>
+					<nav className="flex w-full gap-4" aria-label="Main navigation">
+						<ul className="flex gap-4 overflow-x-auto whitespace-nowrap lg:gap-8 lg:px-0">
+							<NavLink href="/products">All</NavLink>
 							{navLinks.menu?.items?.map((item) => {
 								if (item.category) {
 									return (
-										<ActiveLink key={item.id} href={`/categories/${item.category.slug}`}>
+										<NavLink key={item.id} href={`/categories/${item.category.slug}`}>
 											{item.category.name}
-										</ActiveLink>
+										</NavLink>
 									);
 								}
 								if (item.collection) {
 									return (
-										<ActiveLink key={item.id} href={`/collections/${item.collection.slug}`}>
+										<NavLink key={item.id} href={`/collections/${item.collection.slug}`}>
 											{item.collection.name}
-										</ActiveLink>
+										</NavLink>
 									);
 								}
 								if (item.page) {
 									return (
-										<ActiveLink key={item.id} href={`/pages/${item.page.slug}`}>
+										<NavLink key={item.id} href={`/pages/${item.page.slug}`}>
 											{item.page.title}
-										</ActiveLink>
+										</NavLink>
 									);
 								}
 								if (item.url) {
@@ -55,16 +55,16 @@ export async function Header() {
 								}
 								return null;
 							})}
+						</ul>
+						<div className="ml-auto flex items-center justify-center whitespace-nowrap">
+							<AccountLink />
 						</div>
-					</div>
-					<div className="ml-auto flex items-center justify-center whitespace-nowrap">
-						<AccountLink />
-					</div>
-					<div className="flex items-center">
-						<Suspense fallback={<div className="w-12" />}>
-							<CartNavItem />
-						</Suspense>
-					</div>
+						<div className="flex items-center">
+							<Suspense fallback={<div className="w-12" />}>
+								<CartNavItem />
+							</Suspense>
+						</div>
+					</nav>
 				</div>
 			</div>
 		</header>
